requirements.txt
apache-airflow-providers-amazon
apache-airflow-providers-snowflake
apache-airflow-providers-apache-spark
apache-airflow-providers-databricks
pyspark
dbt-core=1.8.0
dbt-snowflake=1.8.0
boto3
astronomer-cosmos